refactor(launch): one prompt IR for bind refresh; ChildEnvContext projects ResolvedContext
B: ContextEnvRefreshPlan mirrored ComposedLaunchContent field-for-field and
was rebuilt at bind just to swap one guidance block. Delete it and the
from/to_composed cloning; PreparedLaunchContent carries the
ComposedLaunchContent directly, and bind reprojection does a single
dataclasses.replace on the context-env block.

C: ChildEnvContext re-derived work_id/work_dir/context_dirs after calling
ResolvedContext.from_environment — a second scope-precedence engine. Make it
a thin projection of one ResolvedContext; keep only the child-specific rule
(bind ambient work dir to the concrete child_spawn_id).

Tests: split the 870-line test_launch_context_env into contract-scoped files
asserting stable outcomes (work-dir parity, parent dir gone, task cwd) rather
than internal field presence.
